ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.10.0
pydantic-settings==2.6.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.30
numpy<2
httpx==0.25.2
python-dotenv==1.0.0
aiofiles==23.2.1